feat: modify Agent + DDGP Algorithm
Modification to increase the depth of the "Agent" and the "Critic".

With this modifications we reach an average value of 29,X and some
individual values above 30,X. Though not enough to be able to en up
learning.

Maybe a reduction in the initial "action- noise" and tweaks on the
learning rate would help to reach the desired plateau